removing a file, cleaning scat1d, adding doc strings so that it goes straight in the module!
